+++ codersdk/workspacesdk/connector.go
@@ -2,15 +2,9 @@ package workspacesdk
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"fmt"
 	"io"
-	"net"
 	"net/http"
-	"net/http/cookiejar"
-	"net/netip"
-	"strconv"
 	"sync"
 	"time"
 
@@ -25,158 +19,6 @@ import (
 	"github.com/coder/coder/v2/tailnet/proto"
 	"github.com/coder/retry"
 )
-
-type Client struct {
-	client *codersdk.Client
-}
-
-func NewClient(c *codersdk.Client) *Client {
-	return &Client{client: c}
-}
-
-// AgentConnectionInfo returns required information for establishing
-// a connection with a workspace.
-// @typescript-ignore AgentConnectionInfo
-type AgentConnectionInfo struct {
-	DERPMap                  *tailcfg.DERPMap `json:"derp_map"`
-	DERPForceWebSockets      bool             `json:"derp_force_websockets"`
-	DisableDirectConnections bool             `json:"disable_direct_connections"`
-}
-
-func (c *Client) AgentConnectionInfoGeneric(ctx context.Context) (AgentConnectionInfo, error) {
-	res, err := c.client.Request(ctx, http.MethodGet, "/api/v2/workspaceagents/connection", nil)
-	if err != nil {
-		return AgentConnectionInfo{}, err
-	}
-	defer res.Body.Close()
-	if res.StatusCode != http.StatusOK {
-		return AgentConnectionInfo{}, codersdk.ReadBodyAsError(res)
-	}
-
-	var connInfo AgentConnectionInfo
-	return connInfo, json.NewDecoder(res.Body).Decode(&connInfo)
-}
-
-func (c *Client) AgentConnectionInfo(ctx context.Context, agentID uuid.UUID) (AgentConnectionInfo, error) {
-	res, err := c.client.Request(ctx, http.MethodGet, fmt.Sprintf("/api/v2/workspaceagents/%s/connection", agentID), nil)
-	if err != nil {
-		return AgentConnectionInfo{}, err
-	}
-	defer res.Body.Close()
-	if res.StatusCode != http.StatusOK {
-		return AgentConnectionInfo{}, codersdk.ReadBodyAsError(res)
-	}
-
-	var connInfo AgentConnectionInfo
-	return connInfo, json.NewDecoder(res.Body).Decode(&connInfo)
-}
-
-// @typescript-ignore DialAgentOptions
-type DialAgentOptions struct {
-	Logger slog.Logger
-	// BlockEndpoints forced a direct connection through DERP. The Client may
-	// have DisableDirect set which will override this value.
-	BlockEndpoints bool
-}
-
-func (c *Client) DialAgent(dialCtx context.Context, agentID uuid.UUID, options *DialAgentOptions) (agentConn *AgentConn, err error) {
-	if options == nil {
-		options = &DialAgentOptions{}
-	}
-
-	connInfo, err := c.AgentConnectionInfo(dialCtx, agentID)
-	if err != nil {
-		return nil, xerrors.Errorf("get connection info: %w", err)
-	}
-	if connInfo.DisableDirectConnections {
-		options.BlockEndpoints = true
-	}
-
-	ip := tailnet.IP()
-	var header http.Header
-	if headerTransport, ok := c.client.HTTPClient.Transport.(*codersdk.HeaderTransport); ok {
-		header = headerTransport.Header
-	}
-	conn, err := tailnet.NewConn(&tailnet.Options{
-		Addresses:           []netip.Prefix{netip.PrefixFrom(ip, 128)},
-		DERPMap:             connInfo.DERPMap,
-		DERPHeader:          &header,
-		DERPForceWebSockets: connInfo.DERPForceWebSockets,
-		Logger:              options.Logger,
-		BlockEndpoints:      c.client.DisableDirectConnections || options.BlockEndpoints,
-	})
-	if err != nil {
-		return nil, xerrors.Errorf("create tailnet: %w", err)
-	}
-	defer func() {
-		if err != nil {
-			_ = conn.Close()
-		}
-	}()
-
-	headers := make(http.Header)
-	tokenHeader := codersdk.SessionTokenHeader
-	if c.client.SessionTokenHeader != "" {
-		tokenHeader = c.client.SessionTokenHeader
-	}
-	headers.Set(tokenHeader, c.client.SessionToken())
-
-	// New context, separate from dialCtx. We don't want to cancel the
-	// connection if dialCtx is canceled.
-	ctx, cancel := context.WithCancel(context.Background())
-	defer func() {
-		if err != nil {
-			cancel()
-		}
-	}()
-
-	coordinateURL, err := c.client.URL.Parse(fmt.Sprintf("/api/v2/workspaceagents/%s/coordinate", agentID))
-	if err != nil {
-		return nil, xerrors.Errorf("parse url: %w", err)
-	}
-	q := coordinateURL.Query()
-	q.Add("version", proto.CurrentVersion.String())
-	coordinateURL.RawQuery = q.Encode()
-
-	connector := runTailnetAPIConnector(ctx, options.Logger,
-		agentID, coordinateURL.String(),
-		&websocket.DialOptions{
-			HTTPClient: c.client.HTTPClient,
-			HTTPHeader: headers,
-			// Need to disable compression to avoid a data-race.
-			CompressionMode: websocket.CompressionDisabled,
-		},
-		conn,
-	)
-	options.Logger.Debug(ctx, "running tailnet API v2+ connector")
-
-	select {
-	case <-dialCtx.Done():
-		return nil, xerrors.Errorf("timed out waiting for coordinator and derp map: %w", dialCtx.Err())
-	case err = <-connector.connected:
-		if err != nil {
-			options.Logger.Error(ctx, "failed to connect to tailnet v2+ API", slog.Error(err))
-			return nil, xerrors.Errorf("start connector: %w", err)
-		}
-		options.Logger.Debug(ctx, "connected to tailnet v2+ API")
-	}
-
-	agentConn = NewAgentConn(conn, AgentConnOptions{
-		AgentID: agentID,
-		CloseFunc: func() error {
-			cancel()
-			<-connector.closed
-			return conn.Close()
-		},
-	})
-
-	if !agentConn.AwaitReachable(dialCtx) {
-		_ = agentConn.Close()
-		return nil, xerrors.Errorf("timed out waiting for agent to become reachable: %w", dialCtx.Err())
-	}
-
-	return agentConn, nil
-}
 
 // tailnetConn is the subset of the tailnet.Conn methods that tailnetAPIConnector uses.  It is
 // included so that we can fake it in testing.
@@ -389,67 +231,4 @@ func (tac *tailnetAPIConnector) derpMap(client proto.DRPCTailnetClient) error {
 		dm := tailnet.DERPMapFromProto(dmp)
 		tac.conn.SetDERPMap(dm)
 	}
-}
-
-// @typescript-ignore:WorkspaceAgentReconnectingPTYOpts
-type WorkspaceAgentReconnectingPTYOpts struct {
-	AgentID   uuid.UUID
-	Reconnect uuid.UUID
-	Width     uint16
-	Height    uint16
-	Command   string
-
-	// SignedToken is an optional signed token from the
-	// issue-reconnecting-pty-signed-token endpoint. If set, the session token
-	// on the client will not be sent.
-	SignedToken string
-}
-
-// AgentReconnectingPTY spawns a PTY that reconnects using the token provided.
-// It communicates using `agent.ReconnectingPTYRequest` marshaled as JSON.
-// Responses are PTY output that can be rendered.
-func (c *Client) AgentReconnectingPTY(ctx context.Context, opts WorkspaceAgentReconnectingPTYOpts) (net.Conn, error) {
-	serverURL, err := c.client.URL.Parse(fmt.Sprintf("/api/v2/workspaceagents/%s/pty", opts.AgentID))
-	if err != nil {
-		return nil, xerrors.Errorf("parse url: %w", err)
-	}
-	q := serverURL.Query()
-	q.Set("reconnect", opts.Reconnect.String())
-	q.Set("width", strconv.Itoa(int(opts.Width)))
-	q.Set("height", strconv.Itoa(int(opts.Height)))
-	q.Set("command", opts.Command)
-	// If we're using a signed token, set the query parameter.
-	if opts.SignedToken != "" {
-		q.Set(codersdk.SignedAppTokenQueryParameter, opts.SignedToken)
-	}
-	serverURL.RawQuery = q.Encode()
-
-	// If we're not using a signed token, we need to set the session token as a
-	// cookie.
-	httpClient := c.client.HTTPClient
-	if opts.SignedToken == "" {
-		jar, err := cookiejar.New(nil)
-		if err != nil {
-			return nil, xerrors.Errorf("create cookie jar: %w", err)
-		}
-		jar.SetCookies(serverURL, []*http.Cookie{{
-			Name:  codersdk.SessionTokenCookie,
-			Value: c.client.SessionToken(),
-		}})
-		httpClient = &http.Client{
-			Jar:       jar,
-			Transport: c.client.HTTPClient.Transport,
-		}
-	}
-	//nolint:bodyclose
-	conn, res, err := websocket.Dial(ctx, serverURL.String(), &websocket.DialOptions{
-		HTTPClient: httpClient,
-	})
-	if err != nil {
-		if res == nil {
-			return nil, err
-		}
-		return nil, codersdk.ReadBodyAsError(res)
-	}
-	return websocket.NetConn(context.Background(), conn, websocket.MessageBinary), nil
 }
