+++ enterprise/coderd/replicas_test.go
@@ -83,7 +83,7 @@ func TestReplicas(t *testing.T) {
 		require.Len(t, replicas, 2)
 
 		r := setupWorkspaceAgent(t, firstClient, firstUser, 0)
-		conn, err := workspacesdk.NewClient(secondClient).
+		conn, err := workspacesdk.New(secondClient).
 			DialAgent(context.Background(), r.sdkAgent.ID, &workspacesdk.DialAgentOptions{
 				BlockEndpoints: true,
 				Logger:         slogtest.Make(t, nil).Leveled(slog.LevelDebug),
@@ -130,7 +130,7 @@ func TestReplicas(t *testing.T) {
 		require.Len(t, replicas, 2)
 
 		r := setupWorkspaceAgent(t, firstClient, firstUser, 0)
-		conn, err := workspacesdk.NewClient(secondClient).
+		conn, err := workspacesdk.New(secondClient).
 			DialAgent(context.Background(), r.sdkAgent.ID, &workspacesdk.DialAgentOptions{
 				BlockEndpoints: true,
 				Logger:         slogtest.Make(t, nil).Named("client").Leveled(slog.LevelDebug),
