+++ enterprise/wsproxy/wsproxysdk/wsproxysdk.go
@@ -78,7 +78,7 @@ func (c *Client) RequestIgnoreRedirects(ctx context.Context, method, path string
 // DialWorkspaceAgent calls the underlying codersdk.Client's DialWorkspaceAgent
 // method.
 func (c *Client) DialWorkspaceAgent(ctx context.Context, agentID uuid.UUID, options *workspacesdk.DialAgentOptions) (agentConn *workspacesdk.AgentConn, err error) {
-	return workspacesdk.NewClient(c.SDKClient).DialAgent(ctx, agentID, options)
+	return workspacesdk.New(c.SDKClient).DialAgent(ctx, agentID, options)
 }
 
 type IssueSignedAppTokenResponse struct {
