+++ support/support.go
@@ -112,7 +112,7 @@ func DeploymentInfo(ctx context.Context, client *codersdk.Client, log slog.Logge
 	})
 
 	eg.Go(func() error {
-		hr, err := healthsdk.NewHealthClient(client).DebugHealth(ctx)
+		hr, err := healthsdk.New(client).DebugHealth(ctx)
 		if err != nil {
 			return xerrors.Errorf("fetch health report: %w", err)
 		}
@@ -175,7 +175,7 @@ func NetworkInfo(ctx context.Context, client *codersdk.Client, log slog.Logger, 
 			log.Warn(ctx, "agent id required for agent connection info")
 			return nil
 		}
-		connInfo, err := workspacesdk.NewClient(client).AgentConnectionInfo(ctx, agentID)
+		connInfo, err := workspacesdk.New(client).AgentConnectionInfo(ctx, agentID)
 		if err != nil {
 			return xerrors.Errorf("fetch agent conn info: %w", err)
 		}
@@ -333,7 +333,7 @@ func AgentInfo(ctx context.Context, client *codersdk.Client, log slog.Logger, ag
 }
 
 func connectedAgentInfo(ctx context.Context, client *codersdk.Client, log slog.Logger, agentID uuid.UUID, eg *errgroup.Group, a *Agent) (closer func()) {
-	conn, err := workspacesdk.NewClient(client).
+	conn, err := workspacesdk.New(client).
 		DialAgent(ctx, agentID, &workspacesdk.DialAgentOptions{
 			Logger:         log.Named("dial-agent"),
 			BlockEndpoints: false,
